Rename Suffix_info to SuffixInfo, and re-organize the variables in SuffixInfo.py.
