+++ Download_SHM_SPT.py
@@ -2,7 +2,7 @@ import os
 import sys
 from pdart.astroquery.Astroquery import MastSlice, CustomizedQueryMastSlice, ProductSet
 from pdart.pipeline.Directories import DevDirectories
-from pdart.pipeline.Suffix_info import (  # type: ignore
+from pdart.pipeline.SuffixInfo import (  # type: ignore
     ACCEPTED_SUFFIXES,
     PART_OF_ACCEPTED_SUFFIXES,
 )
