+++ pdart/astroquery/Astroquery.py
@@ -10,7 +10,7 @@ from pdart.astroquery.Utils import (
     ymd_tuple_to_mjd,
 )
 
-from pdart.pipeline.Suffix_info import (  # type: ignore
+from pdart.pipeline.SuffixInfo import (  # type: ignore
     ACCEPTED_SUFFIXES,
     PART_OF_ACCEPTED_SUFFIXES,
     ACCEPTED_INSTRUMENTS,
