+++ pdart/db/BundleDB.py
@@ -44,7 +44,7 @@ from pdart.db.SqlAlchTables import (
 )
 from pdart.citations import Citation_Information
 from pdart.db.Utils import file_md5
-from pdart.pipeline.Suffix_info import RAW_SUFFIXES  # type: ignore
+from pdart.pipeline.SuffixInfo import RAW_SUFFIXES  # type: ignore
 from pdart.pds4.HstFilename import HstFilename
 from pdart.pds4.LID import LID
 from pdart.pds4.LIDVID import LIDVID
