+++ pdart/pipeline/CheckDownloads.py
@@ -50,6 +50,7 @@ class CheckDownloads(MarkedStage):
 
         # TODO This might fail if there are no files.  Which might not be
         # a bad thing.
+        print(f"::::::::::mast_downloads_dir: {mast_downloads_dir}")
         assert os.path.isdir(mast_downloads_dir)
 
     def _run(self) -> None:
